Backend/requirements.txt
flask
python-dotenv
flasgger
flask-cors
Flask-SQLAlchemy
Flask-Migrate
flask-socketio
bcrypt
flask-jwt-extended